correct component file name
