+++ Web/starter_project_vanguard/components/blog/AddButton.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Text from '@/components/blog/text'
+import Text from '@/components/blog/Text'
 interface Props {}
 
 const Button: React.FC<Props> = () => {
@@ -7,7 +7,7 @@ const Button: React.FC<Props> = () => {
   return (
     <div className='w-60 '>
       <button
-        className="bg-blue-800 hover:bg-blue-700 text-white py-4 h-16  px-10 rounded-full flex items-center font-montserrat">
+        className="bg-blue-800 hover:bg-blue-700 text-white py-4 h-16  px-10 rounded-full flex items-center font-montserrat btn-pill">
         <i>
             <Text children={'+ New Blog'} size='lg' weight='semibold' family='montserrat'/>
         </i>
