+++ Web/starter_project_vanguard/components/blog/Avatar.tsx
