+++ Web/starter_project_vanguard/components/blog/ProfilePic.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Avatar from './avatar';
+import Avatar from './Avatar';
 
 function ProfilePic() {
   return (
