+++ Web/starter_project_vanguard/components/blog/SearchBar.tsx
